pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pgns-agent-openai"
dynamic = ["version"]
description = "OpenAI Agents SDK adapter for pgns-agent"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [{ name = "pgns" }]
keywords = ["agents", "a2a", "pgns", "openai", "openai-agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "pgns-agent>=0.1",
    "openai-agents>=0.10,<0.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "starlette>=0.41,<1",
]

[project.urls]
Homepage = "https://pgns.io"
Documentation = "https://docs.pgns.io/libraries/pgns-agent"
Repository = "https://github.com/pgns-io/pgns-agent-openai"

[tool.hatch.version]
path = "pgns_agent_openai/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["pgns_agent_openai"]
